+++ R/get_table.R
@@ -6,21 +6,21 @@
 #' 5-year calendar period, race, and sex strata. See `Details` for information on how the
 #' person file must be formatted.
 #'
-#' The person file must contain the variables:
+#' The persondf tibble must contain the variables:
 #' * id,
-#' * gender (numeric: 'M'/'F'),
-#' * race (numeric: 'W'/'N'),
+#' * gender (character: 'M'/'F'),
+#' * race (character: 'W'/'N'),
 #' * dob (date),
 #' * pybegin (date),
 #' * dlo	(date),
-#' * rev (numeric: 5-10),
+#' * rev (numeric: values 5-10),
 #' * code (character: ICD code)
 #'
 #' @param persondf data.frame like object containing one row per person with the required demographic information
-#' @param rateobj a rate object created by the `parseRate` function
-#' @param batch_size a number specifying how many persons to stratify at a time. Default is 25.
+#' @param rateobj a rate object created by the `parseRate` function, or the included rate object `us_119ucod_19602020`
 #' @param strata any additional variables contained in persondf on which to stratify.
 #' Must be wrapped in a `vars()` call from `dplyr`.
+#' @param batch_size a number specifying how many persons to stratify at a time. Default is 500
 #'
 #' @return A data.frame with a row for each strata containing the number of observed
 #' deaths within each of the defined minors/outcomes (`_o1`-`_oxxx`) and the number of person days.
@@ -43,11 +43,25 @@
 #' py_table <- get_table(person, rateobj)
 #'
 #' @importFrom rlang .data
-get_table <- function(persondf, rateobj, batch_size = 25, strata=dplyr::vars()){
+get_table <- function(persondf,
+                      rateobj,
+                      strata=dplyr::vars(),
+                      batch_size = 500){
+
+  get_dates <- function(dob, pybegin, dlo) {
+    full <- sort(c(
+      pybegin,
+      dob + ceiling(rateobj$age_cut * 365.24 - 1),
+      lubridate::mdy(paste0('1/1/', rateobj$cp_cut)),
+      dlo
+    )) %>%
+      unique()
+    return(full[dplyr::between(full, pybegin, dlo)])
+  }
 
   if (length(strata) == 0) ..names <- c() else ..names <- purrr::map_chr(strata, rlang::as_name)
 
-  checkPerson(persondf)
+  checkPerson(persondf, rateobj)
   person_all <- persondf %>%
     dplyr::filter(.data$dlo >= .data$pybegin) #Drop those who died before study begin date
   drop <- nrow(persondf %>%
@@ -86,19 +100,21 @@ get_table <- function(persondf, rateobj, batch_size = 25, strata=dplyr::vars()){
 
     # Get long person file
     person_long <- person %>%
-      dplyr::group_by(.data$id) %>%
-      expand_dates(.data$pybegin, .data$dlo) %>%
+      dplyr::mutate(date = purrr::pmap(list(.data$dob, .data$pybegin, .data$dlo), ~get_dates(..1, ..2, ..3))) %>%
+      tidyr::unnest(date) %>%
       dplyr::mutate(age = floor(as.numeric(.data$date - .data$dob + 1)/365.24),
-             year = lubridate::year(.data$date),
-             ageCat = cut(.data$age, c(-Inf, 3:17 * 5, Inf), right=FALSE), #5-year age
-             CPCat = cut(.data$year, c(-Inf, 380:600 * 5), right=FALSE)) #5-year CP
+                    year = lubridate::year(.data$date),
+                    ageCat = cut(.data$age, c(-Inf, rateobj$age_cut), right=FALSE), #5-year age
+                    CPCat = cut(.data$year, c(-Inf, rateobj$cp_cut), right=FALSE)) %>%  #5-year CP
+      dplyr::group_by(.data$id) %>%
+      dplyr::mutate(pdays = difftime(dplyr::lead(.data$date), .data$date, units = 'days') %>% as.numeric(),
+                    pdays = dplyr::if_else(dplyr::row_number() == dplyr::n(), 1L, as.integer(.data$pdays)))
 
 
     # Collapse into categories
-    options(dplyr.summarise.inform = FALSE)
     py_days <- person_long %>%
       dplyr::group_by(.data$ageCat, .data$CPCat, .data$gender, .data$race, !!!strata) %>%
-      dplyr::summarize(pdays = dplyr::n())
+      dplyr::summarize(pdays = sum(pdays))
     py_min <- person_long %>%
       dplyr::filter(.data$dlo == .data$date) %>%
       dplyr::group_by(.data$ageCat, .data$CPCat, .data$gender, .data$race, !!!strata, .data$minor) %>%
@@ -133,7 +149,6 @@ get_table <- function(persondf, rateobj, batch_size = 25, strata=dplyr::vars()){
     dplyr::mutate(ageCat = as.character(.data$ageCat),
            CPCat= as.character(.data$CPCat))
 
-
   # Check all categories are within rate file
   py_table <- checkStrata(py_table, rateobj)[[1]]
 
@@ -142,7 +157,7 @@ get_table <- function(persondf, rateobj, batch_size = 25, strata=dplyr::vars()){
     dplyr::mutate(obs = rowSums(dplyr::select(., dplyr::starts_with('_o')))) %>%
     dplyr::summarize(obs = sum(.data$obs),
               py = sum(.data$pdays)/365.25)
-  message("- Person Table successfully created\n ",
+  message("- Person Table successfully created\n  ",
           round(ct$py,0), ' person-years and ', ct$obs,
           " deaths in final table")