+++ NEWS.md
@@ -1,3 +1,9 @@
+# LTASR 0.1.0   
+
+* Added functions `get_table_history` and `get_table_history_est` adding the ability
+to stratify cohorts by a time-dependent exposure variable.  
+* Edited logic of `get_table` functions to improve speed.  
+
 # LTASR 0.0.1
 
 * Initial commit.